refactor: separate out notification card types from notificationCard
